// frontend/src/api/schema.d.ts
/**
 * This file was auto-generated by openapi-typescript.
 * Do not make direct changes to the file.
 */

export interface paths {
    "/": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["AppController_getHello"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/rsvp": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        post: operations["RsvpController_create"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: {
        /** @enum {string} */
        RsvpStatus: "ATTENDING" | "NOT_ATTENDING";
        /** @enum {string} */
        GuestCategory: "GROOM" | "BRIDE";
        /** @enum {string} */
        AgeCategory: "ADULT" | "CHILD" | "INFANT" | "BABY";
        /** @enum {string} */
        DietaryRestriction: "WITH" | "WITHOUT";
        CreateRsvpGuestItemDto: {
            status: components["schemas"]["RsvpStatus"];
            guestCategory: components["schemas"]["GuestCategory"];
            japaneseLastName: string;
            japaneseFirstName: string;
            kanaLastName: string;
            kanaFirstName: string;
            email?: string;
            zipcode?: string;
            address?: string;
            building?: string;
            phone?: string;
            ageCategory?: components["schemas"]["AgeCategory"];
            dietaryRestrictions: components["schemas"]["DietaryRestriction"];
            allergyInfo?: string;
            afterParty?: boolean;
            message?: string;
        };
        CreateRsvpBatchDto: {
            guests: components["schemas"]["CreateRsvpGuestItemDto"][];
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}
export type $defs = Record<string, never>;
export interface operations {
    AppController_getHello: {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content?: never;
            };
        };
    };
    RsvpController_create: {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody: {
            content: {
                "application/json": components["schemas"]["CreateRsvpBatchDto"];
            };
        };
        responses: {
            201: {
                headers: {
                    [name: string]: unknown;
                };
                content?: never;
            };
        };
    };
}